rename route from additional-details to onboarding
